refactor(daemon): split runtime fallback helpers
